- FIXED Ldata type on 2 bytes - FIXED reading of wifi post message in wifi-transmitter

+++ embedded/common/include/common/wifi.h
@@ -19,6 +19,6 @@ namespace wifi {
     ~wifi();
     
     int get(const char* req, callback * r);
-    int post(const char* req, const uint8_t * data, const int Ldata, callback * r);
+    int post(const char* req, const uint8_t * data, const uint16_t Ldata, callback * r);
   };
 }
