+++ embedded/common/src/wifi_arduino.cpp
@@ -11,7 +11,7 @@ namespace wifi {
     return 0;
   }
   
-  int wifi::post(const char* req, const uint8_t * data, const int Ldata, callback * r) {
+  int wifi::post(const char* req, const uint8_t * data, const uint16_t Ldata, callback * r) {
     return 0;
   }
 }
